enable sbt tags only conditionally with sbt.tags. fixes #478
In order to pass the specification tags to sbt it is necessary to
instantiate a specification object. Unfortunately sbt invokes this
method without having set the proper testing classloader. This can
lead to bugs when the body of the specification expects some
resources to be accessible (like Akka or Play configuration
files)