Add explicit type to implicit val used before definition. (#559)
Implicits that rely on an inferred return type may not be in scope above their point of definition. In this case, the use comes right before the implicit definition and leads to compilation errors in incremental updates.

You can observe the failure by adding a whitespace to the file and trying to compile (after a full build, so the only file to compile is this one).