Remove sbt-buildinfo from root project (#733)
The root project is only an aggregate of other projects and has no sources of its own. However, the buildinfo plugin creates one, that seems to be useless. This can have an impact on compilation time, but also confuses Hydra, and since this seems an oversight, the simplest is to remove it.